Replaced Bitraverse instances in either/Xor/Validated with Copair
