+++ core/src/main/scala/cats/std/either.scala
@@ -3,7 +3,7 @@ package std
 
 trait EitherInstances extends EitherInstances1 {
 
-  implicit def eitherCopair: Copair[Either] =
+  implicit val eitherBitraverse: Copair[Either] =
     new Copair[Either] {
       def fold[A, B, C](f: Either[A, B])(fa: (A) => C, fb: (B) => C): C =
         f match {
@@ -15,26 +15,6 @@ trait EitherInstances extends EitherInstances1 {
       def right[A, B](b: B): Either[A, B] = Right(b)
     }
 
-  implicit val eitherBitraverse: Bitraverse[Either] =
-    new Bitraverse[Either] {
-      def bitraverse[G[_], A, B, C, D](fab: Either[A, B])(f: A => G[C], g: B => G[D])(implicit G: Applicative[G]): G[Either[C, D]] =
-        fab match {
-          case Left(a) => G.map(f(a))(Left(_))
-          case Right(b) => G.map(g(b))(Right(_))
-        }
-
-      def bifoldLeft[A, B, C](fab: Either[A, B], c: C)(f: (C, A) => C, g: (C, B) => C): C =
-        fab match {
-          case Left(a) => f(c, a)
-          case Right(b) => g(c, b)
-        }
-
-      def bifoldRight[A, B, C](fab: Either[A, B], c: Eval[C])(f: (A, Eval[C]) => Eval[C], g: (B, Eval[C]) => Eval[C]): Eval[C] =
-        fab match {
-          case Left(a) => f(a, c)
-          case Right(b) => g(b, c)
-        }
-    }
 
   implicit def eitherInstances[A]: Monad[Either[A, ?]] with Traverse[Either[A, ?]] =
     new Monad[Either[A, ?]] with Traverse[Either[A, ?]] {
@@ -76,6 +56,7 @@ trait EitherInstances extends EitherInstances1 {
 }
 
 private[std] sealed trait EitherInstances1 extends EitherInstances2 {
+
   implicit def eitherPartialOrder[A, B](implicit A: PartialOrder[A], B: PartialOrder[B]): PartialOrder[Either[A, B]] =
     new PartialOrder[Either[A, B]] {
       def partialCompare(x: Either[A, B], y: Either[A, B]): Double = x.fold(
